requirements.txt
Flask==2.3.2
Flask-MySQLdb==1.0.1
pandas==2.0.3
numpy==1.24.4
openpyxl==3.1.2
matplotlib==3.7.2
seaborn==0.13.2
scikit-learn==1.3.0
XlsxWriter==3.1.2
python-dotenv==1.0.0